+++ core/ajax/sonos3.ajax.php
@@ -115,17 +115,17 @@ try {
 
 	if (init('action') == 'getSonos') {
 		if (init('object_id') == '') {
-			$object = object::byId($_SESSION['user']->getOptions('defaultDashboardObject'));
+			$object = jeeObject::byId($_SESSION['user']->getOptions('defaultDashboardObject'));
 		} else {
-			$object = object::byId(init('object_id'));
+			$object = jeeObject::byId(init('object_id'));
 		}
 		if (!is_object($object)) {
-			$object = object::rootObject();
+			$object = jeeObject::rootObject();
 		}
 		$return = array();
 		$return['eqLogics'] = array();
 		if (init('object_id') == '') {
-			foreach (object::all() as $object) {
+			foreach (jeeObject::all() as $object) {
 				foreach ($object->getEqLogic(true, false, 'sonos3') as $sonos) {
 					$return['eqLogics'][] = $sonos->toHtml(init('version'));
 				}
@@ -134,7 +134,7 @@ try {
 			foreach ($object->getEqLogic(true, false, 'sonos3') as $sonos) {
 				$return['eqLogics'][] = $sonos->toHtml(init('version'));
 			}
-			foreach (object::buildTree($object) as $child) {
+			foreach (jeeObject::buildTree($object) as $child) {
 				$sonoss = $child->getEqLogic(true, false, 'sonos3');
 				if (count($sonoss) > 0) {
 					foreach ($sonoss as $sonos) {